made random

+++ src/main.cpp
@@ -2,12 +2,16 @@
 #include <iostream>
 
 int main(){
-    std::array<long double, 2> x;
-    std::array<long double, 2> y;
+    std::array<double, 2> x;
+    std::array<double, 2> y;
     x[1] = -0.3127;
     x[0] = 11.3245;
     y[1] = 0.1100;
     y[0] = 673.6726;
-    
+	//add input
+    //add setting up and checking best global solution
+    //implement working flow
+    //check on real function
+    //add multithreading
     return 0;
 }
